OVE_WAIT_FOREVER UINT32_MAX -> UINT64_MAX; PM sentinel split

+++ bindings/rust/ove/src/error.rs
@@ -65,9 +65,10 @@ pub type Result<T> = core::result::Result<T, Error>;
 /// Timeout duration meaning "wait forever".
 ///
 /// Matches the numeric value of the C-side `OVE_WAIT_FOREVER` macro
-/// (`UINT32_MAX` nanoseconds) so the substrate's sentinel comparison
-/// succeeds via the implicit promotion described in `<ove/types.h>`.
-pub const WAIT_FOREVER: core::time::Duration = core::time::Duration::from_nanos(u32::MAX as u64);
+/// (`UINT64_MAX` nanoseconds).  `time::dur_to_ns` returns this value
+/// verbatim because `Duration::from_nanos(u64::MAX).as_nanos()` is
+/// exactly `u64::MAX`.
+pub const WAIT_FOREVER: core::time::Duration = core::time::Duration::from_nanos(u64::MAX);
 
 impl Error {
     /// Convert a C return code to `Result<()>`.
